Refactor + Generic repository pattern

+++ HovyFridge.Web/HovyFridge.Api/Data/Entity/Fridge.cs
@@ -12,6 +12,6 @@ namespace HovyFridge.Api.Data.Entity
 
         [NotMapped]
         public int ProductsAmount { get; set; }
-        public int? OwnerId { get; set; }
+        public long? OwnerId { get; set; }
     }
 }
